Replace flowbite-svelte with shadcn-svelte
flowbite-svelte's many limitations have made a lot of stuff annoying or impossible. While it did take some work to get the shadcn-svelte components to (mostly) match, it will hopefully be worth it in the long run.

More testing probably needs to be done to confirm that things are still working.